WIP: Make use of strong typing to simplify API user code
SIRITypesSpecialized is a nearly identical copy of SIRITypes except it makes some properties non nullable. This allows automatic handling by TypedJSON and much less boilerplate in caller code.

The issue is that there is now a lot of duplicated code (SIRITypes / SIRITypesSpecialized) so only the specialized types should be kept