Add StatsService
StatsService enables saving anonymous statistics about bot usage.
The STATS_SLOT_COUNT environnement variable enables creating statistics slots to prevent "outliner users" to have too much impact of stats while still preserving privacy. What this means is that a user can only impact their slot and not the other slots, so that you can filter out slots that are outliners regarding the statistic you are interrested in.
While a user is always assigned the same slot, a slot can and should correspond to multiple users, which enables protecting users privacy. The value of STATS_SLOT_COUNT should therefore be not too high (which would augment the risks of a single slot corresponding only to a single user) but not too small either (which would make it easier for outliners to impact the stats).
Setting STATS_SLOT_COUNT to 1 guarantees a single slot is used for all users but in return doesn't protect anymore against outliners.