Use release command
This changes the Dockerfile and fly.toml file to enable not deploying on build.
This is required in some situations where the build process may not have access
to the environment variables in order to properly call the Discord API.

What we are doing here is that we enable using the DEPLOY_COMMANDS Docker
argument to control wether the build process should deploy or not.
Currently, DEPLOY_COMMANDS is set to YES by default, but this may change
in future releases so it's better to start specifying it explicitly.

The fly.toml file has been changed to include a release_command that
calls the deploy script.

It was important that the ENTRYPOINT was just node, because fly.io uses
the entrypoint to run the release_command and so anything specified
along the ENTRYPOINT would also be run (so we don't want to start the bot
from release_command, we just want to deploy).